New discoveries
So apparently my data is all wrong. So what I need to do the next day is to remake the data for the KNN algorithm in a way where the x variable is all the factors that need to be taken into consideration for the y variable, which is the ranking. This must be done by manually making a list for each attribute and pickling it. There is probably a better way of doing it (and there definitely is), but for now, brute force will do. I'll make it efficient later.