Playlists
Figured out a way to implement playlists and extrapolate data from that. I figured out how to get that data already, but now I need to get the ID for each of the songs, which is a lot more challenging. Need to find a way to do that. But other than this, if KNN doesn't work again, just use the playlists and use the recommendation thing that Spotipy gives me.