Tkinter
I figured out tkinter. The next big step is to fit an image into the text box that will be displayed and how to cycle through the list of ids that I have. The next step after that is to automate the recommendations and make sure that there are no duplicates that are displayed to the user whatsoever. After that, try to see if I can use another user to do these things.